[fix]送信処理部分を切り出し
